python/requirements.txt
# Core LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-chroma==0.1.0
langchain-community==0.0.10

# Vector database
chromadb==0.4.18

# Evaluation frameworks
langsmith==0.0.69
deepeval==0.20.0

# Environment and utilities
python-dotenv==1.0.0
pandas==2.1.4

# Testing
pytest==7.4.3

# Optional: OpenRouter support
openai==1.3.7
